+++ instrumentation/database/sql/splunksql/dbsystem.go
@@ -16,7 +16,7 @@ package splunksql
 
 import (
 	"go.opentelemetry.io/otel/attribute"
-	semconv "go.opentelemetry.io/otel/semconv/v1.7.0"
+	semconv "go.opentelemetry.io/otel/semconv/v1.17.0"
 )
 
 // DBSystem is a database system type.