+++ instrumentation/database/sql/splunksql/stmt.go
@@ -18,7 +18,7 @@ import (
 	"context"
 	"database/sql/driver"
 
-	semconv "go.opentelemetry.io/otel/semconv/v1.7.0"
+	semconv "go.opentelemetry.io/otel/semconv/v1.17.0"
 	"go.opentelemetry.io/otel/trace"
 
 	"github.com/signalfx/splunk-otel-go/instrumentation/database/sql/splunksql/internal/moniker"