+++ instrumentation/github.com/lib/pq/splunkpq/sql.go
@@ -41,6 +41,7 @@ package splunkpq
 
 import (
 	"fmt"
+	"net"
 	"sort"
 	"strconv"
 	"strings"
@@ -76,9 +77,17 @@ func DSNParser(dataSourceName string) (splunksql.ConnectionConfig, error) {
 		connCfg.Host = "localhost"
 	}
 	if strings.HasPrefix(connCfg.Host, "/") {
-		connCfg.NetTransport = splunksql.NetTransportUnix
+		connCfg.NetTransport = splunksql.NetTransportPipe
+		connCfg.NetSockFamily = splunksql.NetSockFamilyUnix
 	} else {
 		connCfg.NetTransport = splunksql.NetTransportTCP
+		if ip := net.ParseIP(connCfg.Host); ip != nil {
+			if ip.To4() != nil {
+				connCfg.NetSockFamily = splunksql.NetSockFamilyInet
+			} else {
+				connCfg.NetSockFamily = splunksql.NetSockFamilyInet6
+			}
+		}
 	}
 	if pInt, err := strconv.Atoi(vals["port"]); err == nil {
 		connCfg.Port = pInt